% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backups.R
\name{backup_csv}
\alias{backup_csv}
\title{Backup only .csv files}
\usage{
backup_csv(source_dir, dest_dir)
}
\arguments{
\item{source_dir}{A string refering to a source directory to browse}

\item{dest_dir}{A string refering to destination. Using only drive letter is not recommended}
}
\value{
nothing
}
\description{
Search given directory recursively and copy all .csv files to destination
}
\examples{
backup_csv("D:/important_data/", "E:/backup")
}
