% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backups.R
\name{backup_data_files}
\alias{backup_data_files}
\title{Backup all data files}
\usage{
backup_data_files(source_dir, dest_dir)
}
\arguments{
\item{source_dir}{A string refering to a source directory to browse}

\item{dest_dir}{A string refering to destination. Using only drive letter is not recommended}
}
\value{
nothing
}
\description{
Search given directory for csv, R, imj, and excell files to copy to destination
}
\examples{
backup_csv("D:/important_data/", "E:/backup")
}
