% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{bplot}
\alias{bplot}
\title{Most basic ggplot}
\usage{
bplot(df, val, var = "treat")
}
\arguments{
\item{df}{Data frame used for plotting}

\item{val}{Name of the column that will be passed to the aes_string as y}

\item{var}{Name of the column that will be passed to the aes_string as x}
}
\value{
Nice and simple plot
}
\description{
The most basic ggplot useful for visualisation of non-complicated data frames
}
\examples{
bplot(mtcars,"cyl","hp")
}
