% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwtools-package.R
\docType{package}
\name{jwtools-package}
\alias{jwtools}
\alias{jwtools-package}
\title{jwtools: Set of useful tools for common problems}
\description{
Offers useful tools for common problems
}
\author{
\strong{Maintainer}: Jan Weslawski \email{rocket_racer@wp.pl}

}
\keyword{internal}
