% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basc_funs.R
\name{new_df}
\alias{new_df}
\title{Initialize data.frame}
\usage{
new_df(x = 1, y = 0)
}
\arguments{
\item{x}{number of columns (1 by default)}

\item{y}{number of rows (0 by default)}
}
\value{
Fresh data.frame with speicified dimensions
}
\description{
Quickly create data frame based on a matrix with specified number of row and cols
}
\examples{
my_df <- new_df(5,10)
}
