% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nlsLM_MC}
\alias{nlsLM_MC}
\title{Monte Carlo simulations of linear least squares models fitted with nlsLM
from the minpack.lm package}
\usage{
nlsLM_MC(model, runs = 1000)
}
\arguments{
\item{model}{A nlsLM object to perform MC calculations with.}

\item{runs}{Number of MC runs to perform (defaults to 1000).}
}
\description{
Perform Monte Carlo (MC) simulations of non linear least squares models
fitted with \code{nlsLM} from the \code{minpack.lm} package
}
\examples{
n <- nls(mpg ~ k * e ^ wt, data = mtcars, start = list(k = 1, e = 2))
nls_MC(n)
}
