% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basc_funs.R
\name{p}
\alias{p}
\title{Quickly paste vector into single string}
\usage{
p(...)
}
\arguments{
\item{...}{Any type of variables that will be pasted into string}
}
\value{
A single string made of the input vector
}
\description{
This function does pretty much the same as paste0, though it is much quicker to call
}
\examples{
var <- "file"
output1 <- p("data/",var,".png")
output2 <- p("cell_",1,"slice_",2)
}
