% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_kinetics_plate}
\alias{plot_kinetics_plate}
\title{Title}
\usage{
plot_kinetics_plate(
  data,
  plate_type = "96",
  path = "raw_plots",
  width = 400,
  height = 300
)
}
\description{
Description
}
