% Generated by roxygen2: do not edit by hand
% Please edit R/qs_workspace.R instead
\name{qs_load_workspace}
\alias{qs_load_workspace}
\title{从 .qs2 文件恢复所有变量到当前环境}
\usage{
qs_load_workspace(file, nthreads = 6, envir = .GlobalEnv, verbose = TRUE)
}
\arguments{
\item{file}{要读取的 \code{.qs2} 文件路径。}

\item{nthreads}{读取线程数，默认 6。}

\item{envir}{变量要恢复到哪个环境，默认 \code{.GlobalEnv}。}

\item{verbose}{是否打印恢复了哪些变量。默认 \code{TRUE}。}
}
\value{
不可见地返回被恢复的变量名字符向量（\code{invisible()}）。
}
\description{
对应你原来用的 \code{load("xxx.RData")}，用来读回
\code{\link{qs_save_workspace}} 保存的多变量文件。
}
\examples{
\dontrun{
qs_load_workspace("core_WT_YMO_workspace.qs2", nthreads = 14)
}
}
