% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biokine.R
\name{read_biokine}
\alias{read_biokine}
\title{Function to import bio-kine ascii files (.bka)}
\usage{
read_biokine(string = "foo")
}
\arguments{
\item{string}{String to count letters in. Defaults to "foo"}
}
\description{
This function allows you to import bio-kine ascii files (.bka).
These are the standard output of BioLogic stopped flow devices.
}
\examples{
read_biokin()
}
