% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biokine.R
\name{read_biokine_all}
\alias{read_biokine_all}
\title{Import all bio-kine ascii files (.bka) in a directory}
\usage{
read_biokine_all(path = ".")
}
\arguments{
\item{path}{Path to the input directory containing the .bka files to read}
}
\description{
This function allows you to import multiple bio-kine ascii files (.bka) in a directory.
These are the standard output of BioLogic stopped flow devices.
}
\examples{
read_biokine_all(path = "my_dir")
}
