% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_fmd_scores}
\alias{read_fmd_scores}
\title{Read fixed width tabular file}
\usage{
read_fmd_scores(path)
}
\arguments{
\item{path}{Path to the input fixed width file.}
}
\description{
This function reads fixed with tabular files like scorefiles
from Rosetta or derived tools as FastMPNNDesign.
}
\examples{
read_fmd_scores(path = "my_file.txt")
}
