% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_kinetic_TECAN}
\alias{read_kinetic_TECAN}
\title{Read TECAN kinetics}
\usage{
read_kinetic_TECAN(
  TECAN_files,
  layout_files,
  n_cond = 3,
  plate_type = "96",
  pre_trigger = 0,
  delay = 0
)
}
\arguments{
\item{TECAN_files}{A vector of TECAN xlsx kinetic files to import.}

\item{layout_files}{A vector of the respective plate layout files.}

\item{n_cond}{Number of conditions per well in the plate layout files.}

\item{plate_type}{Type of plates used. Has to be either "96" or "384".
(default is "96")}

\item{pre_trigger}{Number of pre-trigger time points to discard.}

\item{delay}{Delay time between starting the kinetic and the first time point
(after pre-trigger) in seconds. Can be either a single value which is used
for all plates, or a vector with the same length as \code{TECAN_files} and
\code{layout_files} to specify different delay times for each plate.}
}
\description{
\code{read_kinetic_TECAN()} takes a vector of TECAN xlsx kinetic files and vector
of plate layout files (as xlsx), imports data and adds the information
from the plate layouts.
}
\examples{
read_kinetic_TECAN(TECAN_files = list.files("raw_data", pattern = ".xlsx", full.names = T),
                   layout_files = list.files("plate_layouts", pattern = ".xlsx", full.names = T), 
                   pre_trigger = 2, delay = c(12,24))
}
