% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_proportion_analysis.R
\name{save_dual}
\alias{save_dual}
\title{双格式保存 ggplot 对象（PDF + PNG）}
\usage{
save_dual(filename_prefix, plot, width, height, dpi = 300, bg = "white")
}
\arguments{
\item{filename_prefix}{不含扩展名的文件名前缀}

\item{plot}{ggplot / patchwork 对象}

\item{width, height}{图形尺寸（inch）}

\item{dpi}{PNG分辨率，默认300}

\item{bg}{背景色，默认white}
}
\description{
双格式保存 ggplot 对象（PDF + PNG）
}
