% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dual_gg.R
\name{save_dual_gg}
\alias{save_dual_gg}
\title{保存 ggplot 对象为 PDF + PNG 双格式}
\usage{
save_dual_gg(plot_obj, filename_base, width = 8, height = 6, dpi = 300)
}
\arguments{
\item{plot_obj}{ggplot 对象}

\item{filename_base}{不含后缀的输出路径前缀}

\item{width}{图宽 (inch)}

\item{height}{图高 (inch)}

\item{dpi}{PNG分辨率}
}
\value{
invisible(TRUE/FALSE)，标记保存是否成功
}
\description{
保存 ggplot 对象为 PDF + PNG 双格式
}
