% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{shw}
\alias{shw}
\title{Check normality}
\usage{
shw(df, y, x = "treat")
}
\arguments{
\item{df}{A data frame that contains measured values}

\item{y}{A string containng the named of the column that stores the number values}

\item{x}{A string conting the name of the grouping variable}
}
\value{
Table with info about normality
}
\description{
Perform quick Shapiro-Wilk test of normality on given data frame
}
\examples{
shw(iris, "Sepal.Width", "Species")
}
