% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarise_nls_MC}
\alias{summarise_nls_MC}
\title{Summarize Monte Carlo simulations of linear least squares models}
\usage{
summarise_nls_MC(data, conf.level = 0.95)
}
\arguments{
\item{data}{A tibble generated by nls_MC()}

\item{conf.level}{Level for confidence interval calculation (defaults to 0.95)}
}
\description{
Summarizes the output of the nls_MC() function. For each fitted parameter
mean, median, standard deviation and confidence intervals are computed.
}
\examples{
n <- nls(mpg ~ k * e ^ wt, data = mtcars, start = list(k = 1, e = 2))
s <- nls_MC(n)
summarise_nls_MC(s, conf.level = 0.99)
}
