% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{timeseries_av}
\alias{timeseries_av}
\title{Reduce number of time points by averaging values of multiple time points}
\usage{
timeseries_av(data, window = 5)
}
\arguments{
\item{data}{A data frame with two columns: time and value}

\item{width}{Number of time/value pairs that will be averaged into one pair}
}
\description{
This function reduces the number of time points in a data frame by averaging
values from multiple time points.
}
\examples{
timeseries_av(data = tibble(time = 1:100, value = rnorm(100)), window = 5)
}
