% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basc_funs.R
\name{unsp}
\alias{unsp}
\title{Split and unlist}
\usage{
unsp(x, y)
}
\arguments{
\item{x}{A string to be split}

\item{y}{Pattern used for splitting}
}
\value{
Vector of elements of the split string
}
\description{
Quickly split the string on provided pattern, and unlist it to create a vector
}
\examples{
string <- "this_is_my_string"
output <- unsp(string,"_")
}
