% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basc_funs.R
\name{vd}
\alias{vd}
\title{Quickly peek the first rows of data frame}
\usage{
vd(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
Ten first rows of input data frame
}
\description{
This function allows for peeking first 10 rows of data frame and is easy to call
}
\examples{
vd(df)
}
