% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biokine.R
\name{read_biokine}
\alias{read_biokine}
\title{Import bio-kine ascii files (.bka)}
\usage{
read_biokine(path = "")
}
\arguments{
\item{path}{Path to the input .bka file}
}
\description{
This function allows you to import bio-kine ascii files (.bka).
These are the standard output of BioLogic stopped flow devices.
}
\examples{
read_biokine(path = "my_file.bka")
}
